requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
transformers
torch
openai
python-dotenv